+++ src/plugins.js
@@ -18,7 +18,7 @@ import MarkdownPaste from "./plugins/MarkdownPaste";
 import Ellipsis from "./plugins/Ellipsis";
 import Embeds from "./plugins/Embeds";
 import Chrome from "./plugins/Chrome";
-import TableAlign from "./plugins/TableAlign";
+import Table from "./plugins/Table";
 import Nodes from "./nodes.js";
 import Marks from "./marks.js";
 
@@ -70,7 +70,7 @@ const createPlugins = ({ placeholder, getLinkComponent }: *) => {
       typeCell: "table-cell",
       typeContent: "paragraph",
     }),
-    TableAlign(),
+    Table(),
     Prism({
       onlyIn: node => node.type === "code",
       getSyntax: node => node.data.get("language") || "javascript",
